+++ common/src/client/java/io/github/jamalam360/sort_it_out/client/Config.java
@@ -41,7 +41,11 @@ public class Config extends UserPreferences implements ConfigExtensions<Config> 
 	public void afterSave() {
 		// Only sync when in game
 		if (Minecraft.getInstance().level != null) {
-			this.sync();
+			if (SortItOutClient.justReceivedFromServer) {
+				SortItOutClient.justReceivedFromServer = false;
+			} else {
+				this.sync();
+			}
 		}
 	}